refactor:create clients folder
